Move components out of content
